requirements.txt
openai
langchain-community
streamlit
python-dotenv


